\name{cd4}
\docType{data}
\alias{cd4}
\title{Multicenter AIDS Study}
\description{
  CD4 cell counts for subjects over time.
}
\usage{
data(cd4)
}
\format{
  \code{cd4} 
}
\source{
 
  The data was obtained from \url{http://www.lancaster.ac.uk/staff/diggle/APTS-data-sets/CD4_data.txt}.
}
\references{
  Kaslow, R.A., Ostrow, D.G., Detels, R. et al. (1987). "The Multicenter AIDS Cohort Study: Rationale,
  organization and selected characteristics of the participants" American Journal of Epidemiology, 126,
  310-318.
}

\keyword{datasets}

