\name{mscm}
\docType{data}
\alias{mscm}
\title{Mothers Stress Child Morbidity Data}
\description{
  Mother stress and child sickness over period of around 30 days. 
}
\usage{
data(mscm)
}
\format{
  \code{mscm} 
}
\source{

  The data is described in Zeger and Liang (1986). Mothers and infants were recruited by 
  Professor Cheryl Alexander of the John Hopkins University Department of Maternal and Child Health 
  from the Ambulatory Pediatric Health Services Clinic at Baltimore City Hospital. The mothers were 
  asked to keep a diary of their own stress level and child illness. Baseline characteristics were obtained 
  from a preliminary inteview. Robins, Greenland, and Hu (1999) provide an analysis of the causal effect of mothers stress on 
  child illness using this data.
 
  The data was obtained from \url{https://faculty.washington.edu/heagerty/Books/AnalysisLongitudinal/datasets.html}.
}
\references{

    C. S. Alexander and R. Markowitz (1986). "Maternal employment and use of pediatric services". Medical Care, 24:134-147.

   Zeger and Liang (1986). "Longitudinal Data Analysis for Discrete and Continuous Outcomes," Biometrics, 42, 121-130. 


   Robins, Greenland, and Hu (1999). "Estimation of the Causal Effect of a Time-Varying Exposure on the Marginal Mean of a Repeated Binary Outcome". Journal of American Statistical Association. 94, 687-700.

}

\keyword{datasets}
